Generation until person 25, parameterized gamma models generation , vtime bugfix

+++ src/helpers.h
@@ -13,7 +13,7 @@ enum Estado {S,PI,I,R,V};
 
 int damePuerto(std::string nombreDelPuerto);
 
-VTime toVTime(float f);
+VTime toVTime(double f);
 
 BasicMsgValue* me_infectaron();
 BasicMsgValue* te_infecte();
